# improved performance of Branch#makeMove