--- a/shared/src/main/scala/com/mogproject/mogami/core/state/State.scala
+++ b/shared/src/main/scala/com/mogproject/mogami/core/state/State.scala
@@ -297,7 +297,7 @@ case class State(turn: Player = BLACK,
     val newOccs = getUpdatedOccupancy(move)
 
     val hint = StateHint(
-      hash ^ StateHash.getDifference(hand, move),
+      StateHash.getNextStateHash(this, move),
       newOccs._1,
       newOccs._2,
       newOccs._3,