--- a/shared/src/test/scala/com/mogproject/mogami/bench/TestData.scala
+++ b/shared/src/test/scala/com/mogproject/mogami/bench/TestData.scala
@@ -23,4 +23,7 @@ trait TestData {
     " 6b4d+ 2d3c G*4c 4b4c 5d4c+ 3b2b 4d3c 2b1c N*2e 1c1b",
     " G*2b"
   ).mkString
+
+  // https://shogidb2.com/games/2ee6a9a3754841c721e904e5a5295583297d3dbf
+  val recordUsen01: String = "~0.7ku36e7bq0rs9us2jm8uc31u66q1sc83q31s9lo0io8cm0n88821im51q2o63wr1nq7ga2s68la1a49cm1e26ti0wu5oi0e48hq00i6y219087k09k9h4bf872m3t697m4289q82xabg851o7ls42a6ba3xq8h620s66q3t87222jo4ji2aibei1a04nj15i8kwbgq9l832a7d8bg8bes1ry7g8bh89lqbgq8c84be66s5gebok1e24nj1a0bl0bn2bfi3fibf01ai6t0bg092ibo2bt23ki4p0bfa4ji2aibfi3ei56c6lfbg852s7bs3tsboq2o6bok29i4ji3om3ej05i4nj15ibok2ai45x250bei3ei4nj15ibf02aibb23e0bei15ibsibii3ej15ic06b920dzbi4bjibe03ej15ibjm0am5xm3om4om3t4bl4bi47240b4b9mbzm840bd29424un984bck.r"
 }