before:
```
benchGameLoad: different cache

- avg: 6.257s, min: 5.951s, max: 6.627s

benchGameLoad: same cache

- avg: 1.652s, min: 1.637s, max: 1.662s
```

after:
```
benchGameLoad: different cache

- avg: 6.207s, min: 5.881s, max: 6.594s

benchGameLoad: same cache

- avg: 0.803s, min: 0.791s, max: 0.826s
```